restructured the project a bit.
