+++ Engine/graphics/Graphics.cpp
@@ -20,14 +20,12 @@
 ******************************************************************************************/
 #include "MainWindow.h"
 #include "Graphics.h"
-#include "DXErr.h"
-#include "ChiliException.h"
+#include "utilities/DXErr.h"
+#include "utilities/ChiliException.h"
 #include <assert.h>
 #include <string>
 #include <array>
 #include <functional>
-
-#include "mathFunctions.h"
 
 // Ignore the intellisense error "cannot open source file" for .shh files.
 // They will be created during the build sequence before the preprocessor runs.
