+++ Engine/graphics/MainWindow.cpp
@@ -19,9 +19,9 @@
 *	along with The Chili DirectX Framework.  If not, see <http://www.gnu.org/licenses/>.  *
 ******************************************************************************************/
 #include "MainWindow.h"
-#include "Resource.h"
+#include "utilities/Resource.h"
 #include "Graphics.h"
-#include "ChiliException.h"
+#include "utilities/ChiliException.h"
 #include "Game.h"
 #include <assert.h>
