+++ Engine/graphics/Surface.cpp
@@ -19,9 +19,9 @@
 *	along with The Chili DirectX Framework.  If not, see <http://www.gnu.org/licenses/>.  *
 ******************************************************************************************/
 #define FULL_WINTARD
-#include "ChiliWin.h"
+#include "utilities/ChiliWin.h"
 #include "Surface.h"
-#include "ChiliException.h"
+#include "utilities/ChiliException.h"
 namespace Gdiplus
 {
 	using std::min;
