--- Engine/log/engine/criticalLog.txt
@@ -1,6 +1,0 @@
-[2021-02-10 11:15:05.748] [engineCriticalLogger] [critical] this is level 4 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-10 11:15:05.749] [engineCriticalLogger] [critical] this is level 4 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-10 11:55:07.842] [engineCriticalLogger] [critical] this is level 4 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-10 11:55:07.843] [engineCriticalLogger] [critical] this is level 4 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-13 12:40:52.371] [engineCriticalLogger] [critical] this is level 4 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-13 12:40:52.372] [engineCriticalLogger] [critical] this is level 4 of the log and this empyreani is a variable and another rvalue.. 69
