--- Engine/log/engine/errorLog.txt
@@ -1,6 +1,0 @@
-[2021-02-10 11:15:05.748] [engineErrorLogger] [error] this is level 3 of the log and this empyreani is a variable and another rvalue.. p
-[2021-02-10 11:15:05.749] [engineErrorLogger] [error] this is level 3 of the log and this empyreani is a variable and another rvalue.. p
-[2021-02-10 11:55:07.842] [engineErrorLogger] [error] this is level 3 of the log and this empyreani is a variable and another rvalue.. p
-[2021-02-10 11:55:07.843] [engineErrorLogger] [error] this is level 3 of the log and this empyreani is a variable and another rvalue.. p
-[2021-02-13 12:40:52.371] [engineErrorLogger] [error] this is level 3 of the log and this empyreani is a variable and another rvalue.. p
-[2021-02-13 12:40:52.372] [engineErrorLogger] [error] this is level 3 of the log and this empyreani is a variable and another rvalue.. p
