--- Engine/log/engine/log.txt
@@ -1,30 +1,0 @@
-[2021-02-10 11:15:05.747] [engineTraceLogger] [trace] this is level 0 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-10 11:15:05.748] [engineInfoLogger] [info] this is level 1 of the log and this empyreani is a variable and another rvalue.. 255
-[2021-02-10 11:15:05.748] [engineWarnLogger] [warning] this is level 2 of the log and this empyreani is a variable and another rvalue.. 73
-[2021-02-10 11:15:05.748] [engineErrorLogger] [error] this is level 3 of the log and this empyreani is a variable and another rvalue.. p
-[2021-02-10 11:15:05.748] [engineCriticalLogger] [critical] this is level 4 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-10 11:15:05.748] [engineTraceLogger] [trace] this is level 0 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-10 11:15:05.748] [engineInfoLogger] [info] this is level 1 of the log and this empyreani is a variable and another rvalue.. 255
-[2021-02-10 11:15:05.748] [engineWarnLogger] [warning] this is level 2 of the log and this empyreani is a variable and another rvalue.. 73
-[2021-02-10 11:15:05.749] [engineErrorLogger] [error] this is level 3 of the log and this empyreani is a variable and another rvalue.. p
-[2021-02-10 11:15:05.749] [engineCriticalLogger] [critical] this is level 4 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-10 11:55:07.842] [engineTraceLogger] [trace] this is level 0 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-10 11:55:07.842] [engineInfoLogger] [info] this is level 1 of the log and this empyreani is a variable and another rvalue.. 255
-[2021-02-10 11:55:07.842] [engineWarnLogger] [warning] this is level 2 of the log and this empyreani is a variable and another rvalue.. 73
-[2021-02-10 11:55:07.842] [engineErrorLogger] [error] this is level 3 of the log and this empyreani is a variable and another rvalue.. p
-[2021-02-10 11:55:07.842] [engineCriticalLogger] [critical] this is level 4 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-10 11:55:07.842] [engineTraceLogger] [trace] this is level 0 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-10 11:55:07.842] [engineInfoLogger] [info] this is level 1 of the log and this empyreani is a variable and another rvalue.. 255
-[2021-02-10 11:55:07.842] [engineWarnLogger] [warning] this is level 2 of the log and this empyreani is a variable and another rvalue.. 73
-[2021-02-10 11:55:07.843] [engineErrorLogger] [error] this is level 3 of the log and this empyreani is a variable and another rvalue.. p
-[2021-02-10 11:55:07.843] [engineCriticalLogger] [critical] this is level 4 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-13 12:40:52.371] [engineTraceLogger] [trace] this is level 0 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-13 12:40:52.371] [engineInfoLogger] [info] this is level 1 of the log and this empyreani is a variable and another rvalue.. 255
-[2021-02-13 12:40:52.371] [engineWarnLogger] [warning] this is level 2 of the log and this empyreani is a variable and another rvalue.. 73
-[2021-02-13 12:40:52.371] [engineErrorLogger] [error] this is level 3 of the log and this empyreani is a variable and another rvalue.. p
-[2021-02-13 12:40:52.371] [engineCriticalLogger] [critical] this is level 4 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-13 12:40:52.371] [engineTraceLogger] [trace] this is level 0 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-13 12:40:52.371] [engineInfoLogger] [info] this is level 1 of the log and this empyreani is a variable and another rvalue.. 255
-[2021-02-13 12:40:52.372] [engineWarnLogger] [warning] this is level 2 of the log and this empyreani is a variable and another rvalue.. 73
-[2021-02-13 12:40:52.372] [engineErrorLogger] [error] this is level 3 of the log and this empyreani is a variable and another rvalue.. p
-[2021-02-13 12:40:52.372] [engineCriticalLogger] [critical] this is level 4 of the log and this empyreani is a variable and another rvalue.. 69
