--- Engine/log/engine/traceLog.txt
@@ -1,6 +1,0 @@
-[2021-02-10 11:15:05.747] [engineTraceLogger] [trace] this is level 0 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-10 11:15:05.748] [engineTraceLogger] [trace] this is level 0 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-10 11:55:07.842] [engineTraceLogger] [trace] this is level 0 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-10 11:55:07.842] [engineTraceLogger] [trace] this is level 0 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-13 12:40:52.371] [engineTraceLogger] [trace] this is level 0 of the log and this empyreani is a variable and another rvalue.. 69
-[2021-02-13 12:40:52.371] [engineTraceLogger] [trace] this is level 0 of the log and this empyreani is a variable and another rvalue.. 69
