--- Engine/log/engine/warnLog.txt
@@ -1,6 +1,0 @@
-[2021-02-10 11:15:05.748] [engineWarnLogger] [warning] this is level 2 of the log and this empyreani is a variable and another rvalue.. 73
-[2021-02-10 11:15:05.748] [engineWarnLogger] [warning] this is level 2 of the log and this empyreani is a variable and another rvalue.. 73
-[2021-02-10 11:55:07.842] [engineWarnLogger] [warning] this is level 2 of the log and this empyreani is a variable and another rvalue.. 73
-[2021-02-10 11:55:07.842] [engineWarnLogger] [warning] this is level 2 of the log and this empyreani is a variable and another rvalue.. 73
-[2021-02-13 12:40:52.371] [engineWarnLogger] [warning] this is level 2 of the log and this empyreani is a variable and another rvalue.. 73
-[2021-02-13 12:40:52.372] [engineWarnLogger] [warning] this is level 2 of the log and this empyreani is a variable and another rvalue.. 73
