--- Engine/mathFunctions.h
@@ -1,3 +1,0 @@
-#pragma once
-float normalize(float val, float min, float max);
-float interpolateTo(float x, float y, float alpha);
