+++ Engine/utilities/COMInitializer.cpp
