+++ chili_framework.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31005.135
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Engine", "Engine\Engine.vcxproj", "{F9CB43F6-ABF2-4FE8-A5FD-0DFC16D4F4B2}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Release|x64 = Release|x64
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{F9CB43F6-ABF2-4FE8-A5FD-0DFC16D4F4B2}.Debug|x64.ActiveCfg = Debug|x64
+		{F9CB43F6-ABF2-4FE8-A5FD-0DFC16D4F4B2}.Debug|x64.Build.0 = Debug|x64
+		{F9CB43F6-ABF2-4FE8-A5FD-0DFC16D4F4B2}.Release|x64.ActiveCfg = Release|x64
+		{F9CB43F6-ABF2-4FE8-A5FD-0DFC16D4F4B2}.Release|x64.Build.0 = Release|x64
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {0E73B02C-969A-4BD6-979E-9D70ED5AE982}
+	EndGlobalSection
+EndGlobal
